+++ proj/speech_to_text/speaker_identification.py
@@ -1,7 +1,7 @@
 import json
 import os
 from langchain import OpenAI, PromptTemplate
-from utils import check_environment_vars
+from common.checks import check_environment_vars
 
 check_environment_vars()
 OPENAI_API_KEY = os.environ['OPENAI_API_KEY']
@@ -20,6 +20,27 @@ Description:
 Transcript:
 "{transcript}"
 """
+
+def add_speaker_identification(transcript, speakers_json_filename):
+    """
+    Add the speaker identification to the transcript
+    """
+    speakers = transcript['speakers']
+
+    if not os.path.exists(speakers_json_filename):
+        # Do the speaker identification using LLM
+        speakers_dict = run_llm_speaker_identification(transcript)
+        # Save the speakers to the json file
+        with open(speakers_json_filename, 'w') as f:
+            json.dump(speakers_dict, f)
+    else:
+        with open(speakers_json_filename, 'r') as f:
+            speakers_dict = json.load(f)
+
+    speakers = [speakers_dict.get(speaker, "Unknown Speaker") for speaker in speakers]
+    transcript['speakers'] = speakers
+
+    return transcript
 
 
 def run_llm_speaker_identification(transcript):
